refactor(service): move link-preview endpoint to backend service
