+++ apps/service/package.json
@@ -32,6 +32,7 @@
     "graphql": "^16.11.0",
     "hono": "^4.10.3",
     "hono-rate-limiter": "^0.4.2",
+    "jsdom": "^27.0.1",
     "react": "catalog:react19",
     "react-dom": "catalog:react19",
     "resend": "^6.3.0",
@@ -42,6 +43,7 @@
     "@chia/eslint-config": "workspace:*",
     "@chia/tailwind": "workspace:*",
     "@chia/tsconfig": "workspace:*",
+    "@types/jsdom": "^27.0.0",
     "@types/react": "catalog:react19",
     "@types/react-dom": "catalog:react19",
     "eslint": "catalog:",
