+++ apps/service/src/bootstrap.ts
@@ -17,6 +17,7 @@ import emailRoutes from "@/controllers/email.controller";
 import feedsRoutes from "@/controllers/feeds.controller";
 import healthRoutes from "@/controllers/health.controller";
 import spotifyRoutes from "@/controllers/spotify.controller";
+import toolingsRoutes from "@/controllers/toolings.controller";
 import triggerRoutes from "@/controllers/trigger.controller";
 import trpcRoutes from "@/controllers/trpc.controller";
 import { env } from "@/env";
@@ -153,7 +154,9 @@ const bootstrap = <TContext extends HonoContext>(
   app
     .use("/api/v1/trigger", timeout(env.TIMEOUT_MS))
     .route("/api/v1/trigger", triggerRoutes);
-
+  app
+    .use("/api/v1/toolings", timeout(env.TIMEOUT_MS))
+    .route("/api/v1/toolings", toolingsRoutes);
   console.log(
     `Server is running on port ${port}, go to http://localhost:${port}`
   );
