+++ apps/www/package.json
@@ -56,7 +56,6 @@
     "framer-motion": "12.23.24",
     "fumadocs-mdx": "catalog:",
     "fumadocs-ui": "catalog:",
-    "jsdom": "^27.0.1",
     "katex": "^0.16.25",
     "million": "^3.1.11",
     "next": "16.0.1",
@@ -84,7 +83,6 @@
     "@testing-library/jest-dom": "^6.9.1",
     "@testing-library/react": "^16.3.0",
     "@testing-library/user-event": "^14.6.1",
-    "@types/jsdom": "^27.0.0",
     "@types/mdx": "^2.0.13",
     "@types/react": "catalog:react19",
     "@types/react-dom": "catalog:react19",
