--- apps/www/src/app/api/(v1)/v1/link-preview/_utils.ts
@@ -1,5 +1,0 @@
-export {
-  type DocResponse,
-  type PreviewDTO,
-  previewSchema,
-} from "@chia/ui/link";
