+++ packages/ui/src/link.tsx
@@ -10,7 +10,8 @@ import NextLink from "next/link";
 import type { LinkProps as NextLinkProps } from "next/link";
 import * as z from "zod";
 
-import { post, isUrl, handleKyError, isURLInstance } from "@chia/utils";
+import { isUrl, handleKyError, isURLInstance } from "@chia/utils";
+import { serviceRequest } from "@chia/utils";
 
 import { cn } from "../utils/cn.util";
 import { Avatar, AvatarFallback, AvatarImage } from "./avatar";
@@ -42,12 +43,18 @@ type LinkProps =
     } & PreviewProps);
 
 interface NeverPreviewProps {
+  /**
+   * @deprecated
+   */
   endpoint?: never;
   previewContent?: never;
   queryOptions?: never;
 }
 
 interface PreviewProps {
+  /**
+   * @deprecated
+   */
   endpoint?: string;
   children?:
     | ReactNode
@@ -150,7 +157,8 @@ const PreviewCard: FC<LinkProps & { preview: true }> = ({
   previewContent,
   children,
   href,
-  endpoint = "/api/v1/link-preview",
+  // eslint-disable-next-line @typescript-eslint/no-deprecated
+  endpoint: _endpoint,
   preview: _preview,
   experimental: _experimental,
   locale: _locale,
@@ -162,15 +170,14 @@ const PreviewCard: FC<LinkProps & { preview: true }> = ({
     ...queryOptions,
     queryKey: ["link-preview", { href }],
     queryFn: async ({ signal }) => {
-      return await post<DocResponse, PreviewDTO>(
-        endpoint,
-        {
-          href: href.toString(),
-        },
-        {
+      return await serviceRequest()
+        .post("toolings/link-preview", {
+          json: {
+            href: href.toString(),
+          },
           signal,
-        }
-      );
+        })
+        .json<DocResponse>();
     },
     enabled: isOpen && isUrl(href),
   });
